Split Powernet into Powernet and PowernetPowerManager
This restructuring allows for arbitrary networks to be built on the powernet wires.

Still uses the same infrastructure, just moves the code around.

...But really, this is so that solar control computers will be able to find their targets efficiently.
